--- packages/react-native-client/android/src/main/java/io/fishjam/reactnative/managers/CameraDeviceChangedListenerManager.kt
@@ -1,12 +1,0 @@
-package io.fishjam.reactnative.managers
-
-interface LocalCameraTrackChangedListener {
-  fun onLocalCameraTrackChanged()
-}
-
-class LocalCameraTracksChangedListenersManager :
-  ListenerManager<LocalCameraTrackChangedListener>() {
-  fun notifyListeners() {
-    for (listener in listeners) listener.onLocalCameraTrackChanged()
-  }
-}
